abstract couchbase test setup
