+++ src/test/groovy/edu/ie3/datamodel/io/source/couchbase/CouchbaseWeatherSourceCosmoIT.groovy
@@ -5,114 +5,30 @@
  */
 package edu.ie3.datamodel.io.source.couchbase
 
-import edu.ie3.datamodel.io.connectors.CouchbaseConnector
 import edu.ie3.datamodel.io.factory.timeseries.CosmoTimeBasedWeatherValueFactory
 import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries
 import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue
 import edu.ie3.datamodel.models.value.WeatherValue
 import edu.ie3.test.common.CosmoWeatherTestData
-import edu.ie3.test.helper.TestContainerHelper
-import edu.ie3.test.helper.WeatherSourceTestHelper
 import edu.ie3.util.interval.ClosedInterval
 import org.locationtech.jts.geom.Point
-import org.testcontainers.couchbase.BucketDefinition
-import org.testcontainers.couchbase.CouchbaseContainer
 import org.testcontainers.spock.Testcontainers
-import org.testcontainers.utility.MountableFile
-import spock.lang.Shared
-import spock.lang.Specification
-
-import java.time.Duration
 
 @Testcontainers
-class CouchbaseWeatherSourceCosmoIT extends Specification implements TestContainerHelper, WeatherSourceTestHelper {
-
-  @Shared
-  BucketDefinition bucketDefinition = new BucketDefinition("ie3_in")
-
-  @Shared
-  CouchbaseContainer couchbaseContainer = new CouchbaseContainer("couchbase/server:8.0.0")
-  .withBucket(bucketDefinition)
-  .withExposedPorts(8091, 8092, 8093, 8094, 11210)
-  .withStartupAttempts(3) // 3 attempts because startup (node renaming) sometimes fails when executed too early
-
-  @Shared
-  CouchbaseWeatherSource source
-
-  static String coordinateIdColumnName = "coordinateid"
-
-  def setupSpec() {
-    // create an index for the document keys
-    couchbaseContainer.execInContainer("cbq",
-        "-e", "http://localhost:8093",
-        "-u", couchbaseContainer.username,
-        "-p", couchbaseContainer.password,
-        "-s", "CREATE index id_idx ON `" + bucketDefinition.name + "` (META().id);")
-
-    // Create connector to import the data from json document
-    def connector = new CouchbaseConnector(
-        couchbaseContainer.connectionString,
-        bucketDefinition.name,
-        couchbaseContainer.username,
-        couchbaseContainer.password,
-        Duration.ofSeconds(30))
-
-    // Wait for bucket to be ready
-    println "Waiting for Couchbase bucket to be ready..."
-    int maxTries = 10
-    boolean ready = false
-    for (int i = 0; i < maxTries; i++) {
-      try {
-        connector.getSourceFields()
-        ready = true
-        println "Couchbase bucket ready"
-        break
-      } catch (Exception ex) {
-        if (i % 10 == 0) {
-          println "Waiting for bucket... (try ${i+1})"
-        }
-        sleep(1000)
-      }
-    }
-
-    if (!ready) {
-      println "Couchbase bucket did not become ready in time!"
-      System.out.flush()
-      throw new RuntimeException("Couchbase bucket did not become ready in time!")
-    }
-
-    // Insert test data from JSON file
-    println "Inserting test data from JSON file..."
-    def jsonFile = new File("src/test/resources/edu/ie3/datamodel/io/source/couchbase/_weather/cosmo/weather.json")
-    def jsonSlurper = new groovy.json.JsonSlurper()
-    def weatherDocs = jsonSlurper.parse(jsonFile) as List
-    def insertCount = 0
-    weatherDocs.each { doc ->
-      try {
-        def coordinateId = doc["coordinateid"]
-        def time = doc["time"]
-        def key = "weather::${coordinateId}::${time}"
-        connector.persist(key, doc).join()
-        insertCount++
-      } catch (Exception ex) {
-        println "WARNING: Failed to insert document for coordinateid ${doc["coordinateid"]} and time ${doc["time"]}: ${ex.message}"
-      }
-    }
-    println "Inserted ${insertCount}/${weatherDocs.size()} test documents from JSON file"
-
-    def dtfPattern = "yyyy-MM-dd'T'HH:mm:ssxxx"
-    def weatherFactory = new CosmoTimeBasedWeatherValueFactory()
-    source = new CouchbaseWeatherSource(connector, CosmoWeatherTestData.coordinateSource, coordinateIdColumnName, weatherFactory, dtfPattern)
-    println "setupSpec completed"
-    System.out.flush()
+class CouchbaseWeatherSourceCosmoIT extends AbstractCouchbaseWeatherSourceIT {
+  @Override
+  String getJsonResourcePath() {
+    return "src/test/resources/edu/ie3/datamodel/io/source/couchbase/_weather/cosmo/weather.json"
   }
 
-  def "The test container can establish a valid connection"() {
-    when:
-    def connector = new CouchbaseConnector(couchbaseContainer.connectionString, bucketDefinition.name, couchbaseContainer.username, couchbaseContainer.password)
+  @Override
+  Object getWeatherFactory() {
+    return new CosmoTimeBasedWeatherValueFactory()
+  }
 
-    then:
-    connector.connectionValid
+  @Override
+  Object getCoordinateSource() {
+    return CosmoWeatherTestData.coordinateSource
   }
 
   def "A CouchbaseWeatherSource can read and correctly parse a single value for a specific date and coordinate"() {
